+++ src/utils/utils-common.js
@@ -19,4 +19,5 @@ const updateItem = (items, update) => {
   ];
 };
 
+
 export { getRandomInteger, updateItem };
